// contract/src/enum.ts
export enum TX_STATUS {
    UNPAID = "0",
    PENDING = "1",
    TRANSFERRING = "2",
    SUCCESS = "3",
    CANCELED = "4",
    REFUNDING = "5",
    REFUNDED = "6",
}

export enum GAS {
    FT_ON_PURCHASE = 100_000_000_000_000,
    FT_ON_REFUND = 100_000_000_000_000,
    FT_ON_TRANSFER = 100_000_000_000_000,
    FT_ON_REGISTER = 100_000_000_000_000,
    RESOLVE_BUY_TOKEN = 100_000_000_000_000,
    RESOLVE_REGISTER = 100_000_000_000_000,
    RESOLVE_CREATE_ORDER = 100_000_000_000_000,
    RESOLVE_CANCEL_ORDER = 100_000_000_000_000,
}